+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- HomeMyGroupsWidget renamed to HomeGroupsWidget.
+- Make group images expandable [#2474](https://github.com/rokwire/illinois-app/issues/2474).
 
 ## [4.2.7] - 2022-10-21
 ### Changed